delegated searching cleaning logic to helper function
